refactor(gql): separate into individual files
