+++ nuxt/components/PasswordChange.vue
@@ -69,7 +69,7 @@
 <script>
 import { minLength, required } from 'vuelidate/lib/validators'
 
-import { ACCOUNT_PASSWORD_CHANGE } from '~/scripts/apollo'
+import ACCOUNT_PASSWORD_CHANGE_MUTATION from '~/gql/mutation/accountPasswordChange'
 
 export default {
   data() {
@@ -92,7 +92,7 @@ export default {
       this.$v.form.$reset()
       this.$apollo
         .mutate({
-          mutation: ACCOUNT_PASSWORD_CHANGE,
+          mutation: ACCOUNT_PASSWORD_CHANGE_MUTATION,
           variables: {
             passwordCurrent: this.form['current-password'],
             passwordNew: this.form['new-password'],
