+++ nuxt/components/forms/FormSignIn.vue
@@ -77,7 +77,7 @@
 <script>
 import { helpers, minLength, required } from 'vuelidate/lib/validators'
 
-import { AUTHENTICATE_MUTATION } from '~/scripts/apollo'
+import AUTHENTICATE_MUTATION from '~/gql/mutation/authenticate'
 
 const slug = helpers.regex('slug', /^[-A-Za-z0-9]+$/)
