+++ nuxt/gql/mutation/accountPasswordChange.gql
@@ -1,0 +1,7 @@
+mutation($passwordCurrent: String!, $passwordNew: String!) {
+  accountPasswordChange(
+    input: { passwordCurrent: $passwordCurrent, passwordNew: $passwordNew }
+  ) {
+    clientMutationId
+  }
+}
