+++ nuxt/gql/query/allEvents.gql
@@ -1,0 +1,23 @@
+query($username: String, $limit: Int!, $cursor: Cursor) {
+  allEvents(
+    condition: { organizerUsername: $username }
+    first: $limit
+    after: $cursor
+  ) {
+    nodes {
+      name
+      slug
+      visibility
+      organizerUsername
+      description
+      place
+      start
+      end
+      archived
+    }
+    pageInfo {
+      hasNextPage
+      endCursor
+    }
+  }
+}
