+++ nuxt/gql/query/allProfilePictures.gql
@@ -1,0 +1,7 @@
+query allProfilePictures($username: String!) {
+  allProfilePictures(condition: { username: $username }) {
+    nodes {
+      uploadStorageKey
+    }
+  }
+}
