+++ nuxt/gql/query/allUploads.gql
@@ -1,0 +1,17 @@
+query($username: String, $limit: Int!, $cursor: Cursor) {
+  allUploads(
+    condition: { username: $username }
+    first: $limit
+    after: $cursor
+  ) {
+    nodes {
+      id
+      storageKey
+      sizeByte
+    }
+    pageInfo {
+      hasNextPage
+      endCursor
+    }
+  }
+}
