+++ nuxt/gql/query/eventContactFeedbackData.gql
@@ -1,0 +1,31 @@
+query($organizerUsername: String!, $slug: String!) {
+  eventContactFeedbackData(
+    _organizerUsername: $organizerUsername
+    _slug: $slug
+  ) {
+    event {
+      id
+      name
+      slug
+      visibility
+      organizerUsername
+      description
+      place
+      start
+      end
+      archived
+    }
+    contact {
+      id
+      firstName
+      lastName
+      address
+      emailAddress
+    }
+    invitationFeedbackData {
+      id
+      invitationFeedback
+      paperInvitationFeedback
+    }
+  }
+}
