+++ nuxt/pages/event/_username/_event_name.vue
@@ -176,10 +176,8 @@
 <script>
 import VueMarkdown from 'vue-markdown-konishi'
 
-import {
-  EVENT_CONTACT_FEEDBACK_DATA_QUERY,
-  UPDATE_INVITATION_FEEDBACK_DATUM_BY_ID_MUTATION,
-} from '~/scripts/apollo'
+import EVENT_CONTACT_FEEDBACK_DATA_QUERY from '~/gql/query/eventContactFeedbackData'
+import UPDATE_INVITATION_FEEDBACK_DATUM_BY_ID_MUTATION from '~/gql/mutation/updateInvitationFeedbackDatumById'
 
 export default {
   apollo: {
