+++ nuxt/plugins/global.js
@@ -1,6 +1,7 @@
 import { decode } from 'jsonwebtoken'
 
-import { AUTHENTICATE_MUTATION, JWT_REFRESH_MUTATION } from '~/scripts/apollo'
+import AUTHENTICATE_MUTATION from '../gql/mutation/authenticate'
+import JWT_REFRESH_MUTATION from '../gql/mutation/jwtRefresh'
 
 export const EVENT_DESCRIPTION_MAXIMUM = 10000
 export const EVENT_NAME_MAXIMUM = 100
